requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.1.1
python-multipart==0.0.9